Crate rusoto_servicediscovery

AWS Cloud Map lets you configure public DNS, private DNS, or HTTP namespaces that your microservice applications run in. When an instance of the service becomes available, you can call the AWS Cloud Map API to register the instance with AWS Cloud Map. For public or private DNS namespaces, AWS Cloud Map automatically creates DNS records and an optional health check. Clients that submit public or private DNS queries, or HTTP requests, for the service receive an answer that contains up to eight healthy records.

If you're using the service, you're probably looking for ServiceDiscoveryClient and ServiceDiscovery.

Structs

CreateHttpNamespaceRequest

CreateHttpNamespaceResponse

CreatePrivateDnsNamespaceRequest

CreatePrivateDnsNamespaceResponse

CreatePublicDnsNamespaceRequest

CreatePublicDnsNamespaceResponse

CreateServiceRequest

CreateServiceResponse

DeleteNamespaceRequest

DeleteNamespaceResponse

DeleteServiceRequest

DeleteServiceResponse

DeregisterInstanceRequest

DeregisterInstanceResponse

DiscoverInstancesRequest

DiscoverInstancesResponse

DnsConfig
A complex type that contains information about the Amazon Route 53 DNS records that you want AWS Cloud Map to create when you register an instance.

DnsConfigChange
A complex type that contains information about changes to the Route 53 DNS records that AWS Cloud Map creates when you register an instance.

DnsProperties
A complex type that contains the ID for the Route 53 hosted zone that AWS Cloud Map creates when you create a namespace.

DnsRecord
A complex type that contains information about the Route 53 DNS records that you want AWS Cloud Map to create when you register an instance.

GetInstanceRequest

GetInstanceResponse

GetInstancesHealthStatusRequest

GetInstancesHealthStatusResponse

GetNamespaceRequest

GetNamespaceResponse

GetOperationRequest

GetOperationResponse

GetServiceRequest

GetServiceResponse

HealthCheckConfig
Public DNS namespaces only. A complex type that contains settings for an optional health check. If you specify settings for a health check, AWS Cloud Map associates the health check with the records that you specify in DnsConfig.

If you specify a health check configuration, you can specify either HealthCheckCustomConfig or HealthCheckConfig but not both.

Health checks are basic Route 53 health checks that monitor an AWS endpoint. For information about pricing for health checks, see Amazon Route 53 Pricing.

Note the following about configuring health checks.

A and AAAA records

If DnsConfig includes configurations for both A and AAAA records, AWS Cloud Map creates a health check that uses the IPv4 address to check the health of the resource. If the endpoint that is specified by the IPv4 address is unhealthy, Route 53 considers both the A and AAAA records to be unhealthy.

CNAME records

You can't specify settings for HealthCheckConfig when the DNSConfig includes CNAME for the value of Type. If you do, the CreateService request will fail with an InvalidInput error.

Request interval

A Route 53 health checker in each health-checking region sends a health check request to an endpoint every 30 seconds. On average, your endpoint receives a health check request about every two seconds. However, health checkers don't coordinate with one another, so you'll sometimes see several requests per second followed by a few seconds with no health checks at all.

Health checking regions

Health checkers perform checks from all Route 53 health-checking regions. For a list of the current regions, see Regions.

Alias records

When you register an instance, if you include the AWS_ALIAS_DNS_NAME attribute, AWS Cloud Map creates a Route 53 alias record. Note the following:

Route 53 automatically sets EvaluateTargetHealth to true for alias records. When EvaluateTargetHealth is true, the alias record inherits the health of the referenced AWS resource. such as an ELB load balancer. For more information, see EvaluateTargetHealth.

If you include HealthCheckConfig and then use the service to register an instance that creates an alias record, Route 53 doesn't create the health check.

Charges for health checks

Health checks are basic Route 53 health checks that monitor an AWS endpoint. For information about pricing for health checks, see Amazon Route 53 Pricing.

HealthCheckCustomConfig
A complex type that contains information about an optional custom health check. A custom health check, which requires that you use a third-party health checker to evaluate the health of your resources, is useful in the following circumstances:

You can't use a health check that is defined by HealthCheckConfig because the resource isn't available over the internet. For example, you can use a custom health check when the instance is in an Amazon VPC. (To check the health of resources in a VPC, the health checker must also be in the VPC.)

You want to use a third-party health checker regardless of where your resources are.

If you specify a health check configuration, you can specify either HealthCheckCustomConfig or HealthCheckConfig but not both.

To change the status of a custom health check, submit an UpdateInstanceCustomHealthStatus request. Cloud Map doesn't monitor the status of the resource, it just keeps a record of the status specified in the most recent UpdateInstanceCustomHealthStatus request.

Here's how custom health checks work:

You create a service and specify a value for FailureThreshold.

The failure threshold indicates the number of 30-second intervals you want AWS Cloud Map to wait between the time that your application sends an UpdateInstanceCustomHealthStatus request and the time that AWS Cloud Map stops routing internet traffic to the corresponding resource.

You register an instance.

You configure a third-party health checker to monitor the resource that is associated with the new instance.

AWS Cloud Map doesn't check the health of the resource directly.

The third-party health-checker determines that the resource is unhealthy and notifies your application.

Your application submits an UpdateInstanceCustomHealthStatus request.

AWS Cloud Map waits for (FailureThreshold x 30) seconds.

If another UpdateInstanceCustomHealthStatus request doesn't arrive during that time to change the status back to healthy, AWS Cloud Map stops routing traffic to the resource.

Note the following about configuring custom health checks.

HttpInstanceSummary
In a response to a DiscoverInstance request, HttpInstanceSummary contains information about one instance that matches the values that you specified in the request.

HttpProperties
A complex type that contains the name of an HTTP namespace.

Instance
A complex type that contains information about an instance that AWS Cloud Map creates when you submit a RegisterInstance request.

InstanceSummary
A complex type that contains information about the instances that you registered by using a specified service.

ListInstancesRequest

ListInstancesResponse

ListNamespacesRequest

ListNamespacesResponse

ListOperationsRequest

ListOperationsResponse

ListServicesRequest

ListServicesResponse

Namespace
A complex type that contains information about a specified namespace.

NamespaceFilter
A complex type that identifies the namespaces that you want to list. You can choose to list public or private namespaces.

NamespaceProperties
A complex type that contains information that is specific to the namespace type.

NamespaceSummary
A complex type that contains information about a namespace.

Operation
A complex type that contains information about a specified operation.

OperationFilter
A complex type that lets you select the operations that you want to list.

OperationSummary
A complex type that contains information about an operation that matches the criteria that you specified in a ListOperations request.

RegisterInstanceRequest

RegisterInstanceResponse

Service
A complex type that contains information about the specified service.

ServiceChange
A complex type that contains changes to an existing service.

ServiceDiscoveryClient
A client for the ServiceDiscovery API.

ServiceFilter
A complex type that lets you specify the namespaces that you want to list services for.

ServiceSummary
A complex type that contains information about a specified service.

UpdateInstanceCustomHealthStatusRequest

UpdateServiceRequest

UpdateServiceResponse

Enums

CreateHttpNamespaceError
Errors returned by CreateHttpNamespace

CreatePrivateDnsNamespaceError
Errors returned by CreatePrivateDnsNamespace

CreatePublicDnsNamespaceError
Errors returned by CreatePublicDnsNamespace

CreateServiceError
Errors returned by CreateService

DeleteNamespaceError
Errors returned by DeleteNamespace

DeleteServiceError
Errors returned by DeleteService

DeregisterInstanceError
Errors returned by DeregisterInstance

DiscoverInstancesError
Errors returned by DiscoverInstances

GetInstanceError
Errors returned by GetInstance

GetInstancesHealthStatusError
Errors returned by GetInstancesHealthStatus

GetNamespaceError
Errors returned by GetNamespace

GetOperationError
Errors returned by GetOperation

GetServiceError
Errors returned by GetService

ListInstancesError
Errors returned by ListInstances

ListNamespacesError
Errors returned by ListNamespaces

ListOperationsError
Errors returned by ListOperations

ListServicesError
Errors returned by ListServices

RegisterInstanceError
Errors returned by RegisterInstance

UpdateInstanceCustomHealthStatusError
Errors returned by UpdateInstanceCustomHealthStatus

UpdateServiceError
Errors returned by UpdateService

Traits

ServiceDiscovery
Trait representing the capabilities of the ServiceDiscovery API. ServiceDiscovery clients implement this trait.