Crate rusoto_servicediscovery

Amazon Route 53 auto naming lets you configure public or private namespaces that your microservice applications run in. When instances of the service become available, you can call the auto naming API to register the instance, and Route 53 automatically creates up to five DNS records and an optional health check. Clients that submit DNS queries for the service receive an answer that contains up to eight healthy records.

If you're using the service, you're probably looking for ServiceDiscoveryClient and ServiceDiscovery.

Structs

CreatePrivateDnsNamespaceRequest

CreatePrivateDnsNamespaceResponse

CreatePublicDnsNamespaceRequest

CreatePublicDnsNamespaceResponse

CreateServiceRequest

CreateServiceResponse

DeleteNamespaceRequest

DeleteNamespaceResponse

DeleteServiceRequest

DeleteServiceResponse

DeregisterInstanceRequest

DeregisterInstanceResponse

DnsConfig
A complex type that contains information about the records that you want Amazon Route 53 to create when you register an instance.

DnsConfigChange
A complex type that contains information about changes to the records that Route 53 creates when you register an instance.

DnsProperties
A complex type that contains the ID for the hosted zone that Route 53 creates when you create a namespace.

DnsRecord
A complex type that contains information about the records that you want Route 53 to create when you register an instance.

GetInstanceRequest

GetInstanceResponse

GetInstancesHealthStatusRequest

GetInstancesHealthStatusResponse

GetNamespaceRequest

GetNamespaceResponse

GetOperationRequest

GetOperationResponse

GetServiceRequest

GetServiceResponse

HealthCheckConfig
Public DNS namespaces only. A complex type that contains settings for an optional health check. If you specify settings for a health check, Amazon Route 53 associates the health check with all the records that you specify in DnsConfig.

A and AAAA records

If DnsConfig includes configurations for both A and AAAA records, Route 53 creates a health check that uses the IPv4 address to check the health of the resource. If the endpoint that is specified by the IPv4 address is unhealthy, Route 53 considers both the A and AAAA records to be unhealthy.

CNAME records

You can't specify settings for HealthCheckConfig when the DNSConfig includes CNAME for the value of Type. If you do, the CreateService request will fail with an InvalidInput error.

Request interval

The health check uses 30 seconds as the request interval. This is the number of seconds between the time that each Route 53 health checker gets a response from your endpoint and the time that it sends the next health check request. A health checker in each data center around the world sends your endpoint a health check request every 30 seconds. On average, your endpoint receives a health check request about every two seconds. Health checkers in different data centers don't coordinate with one another, so you'll sometimes see several requests per second followed by a few seconds with no health checks at all.

Health checking regions

Health checkers perform checks from all Route 53 health-checking regions. For a list of the current regions, see Regions.

Alias records

When you register an instance, if you include the AWS_ALIAS_DNS_NAME attribute, Route 53 creates an alias record. Note the following:

Route 53 automatically sets EvaluateTargetHealth to true for alias records. When EvaluateTargetHealth is true, the alias record inherits the health of the referenced AWS resource. such as an ELB load balancer. For more information, see EvaluateTargetHealth.

If you include HealthCheckConfig and then use the service to register an instance that creates an alias record, Route 53 doesn't create the health check.

For information about the charges for health checks, see Route 53 Pricing.

HealthCheckCustomConfig

Instance
A complex type that contains information about an instance that Amazon Route 53 creates when you submit a RegisterInstance request.

InstanceSummary
A complex type that contains information about the instances that you registered by using a specified service.

ListInstancesRequest

ListInstancesResponse

ListNamespacesRequest

ListNamespacesResponse

ListOperationsRequest

ListOperationsResponse

ListServicesRequest

ListServicesResponse

Namespace
A complex type that contains information about a specified namespace.

NamespaceFilter
A complex type that identifies the namespaces that you want to list. You can choose to list public or private namespaces.

NamespaceProperties
A complex type that contains information that is specific to the namespace type.

NamespaceSummary
A complex type that contains information about a namespace.

Operation
A complex type that contains information about a specified operation.

OperationFilter
A complex type that lets you select the operations that you want to list.

OperationSummary
A complex type that contains information about an operation that matches the criteria that you specified in a ListOperations request.

RegisterInstanceRequest

RegisterInstanceResponse

Service
A complex type that contains information about the specified service.

ServiceChange
A complex type that contains changes to an existing service.

ServiceDiscoveryClient
A client for the ServiceDiscovery API.

ServiceFilter
A complex type that lets you specify the namespaces that you want to list services for.

ServiceSummary
A complex type that contains information about a specified service.

UpdateInstanceCustomHealthStatusRequest

UpdateServiceRequest

UpdateServiceResponse

Enums

CreatePrivateDnsNamespaceError
Errors returned by CreatePrivateDnsNamespace

CreatePublicDnsNamespaceError
Errors returned by CreatePublicDnsNamespace

CreateServiceError
Errors returned by CreateService

DeleteNamespaceError
Errors returned by DeleteNamespace

DeleteServiceError
Errors returned by DeleteService

DeregisterInstanceError
Errors returned by DeregisterInstance

GetInstanceError
Errors returned by GetInstance

GetInstancesHealthStatusError
Errors returned by GetInstancesHealthStatus

GetNamespaceError
Errors returned by GetNamespace

GetOperationError
Errors returned by GetOperation

GetServiceError
Errors returned by GetService

ListInstancesError
Errors returned by ListInstances

ListNamespacesError
Errors returned by ListNamespaces

ListOperationsError
Errors returned by ListOperations

ListServicesError
Errors returned by ListServices

RegisterInstanceError
Errors returned by RegisterInstance

UpdateInstanceCustomHealthStatusError
Errors returned by UpdateInstanceCustomHealthStatus

UpdateServiceError
Errors returned by UpdateService

Traits

ServiceDiscovery
Trait representing the capabilities of the ServiceDiscovery API. ServiceDiscovery clients implement this trait.