Struct rusoto_servicediscovery::ServiceDiscoveryClient

pub struct ServiceDiscoveryClient { /* fields omitted */ }

A client for the ServiceDiscovery API.

Methods

impl ServiceDiscoveryClient

pub fn new(region: Region) -> ServiceDiscoveryClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> ServiceDiscoveryClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,