Trait rusoto_servicediscovery::ServiceDiscovery

pub trait ServiceDiscovery {
    fn create_private_dns_namespace(
        &self, 
        input: CreatePrivateDnsNamespaceRequest
    ) -> RusotoFuture<CreatePrivateDnsNamespaceResponse, CreatePrivateDnsNamespaceError>;

    fn create_public_dns_namespace(
        &self, 
        input: CreatePublicDnsNamespaceRequest
    ) -> RusotoFuture<CreatePublicDnsNamespaceResponse, CreatePublicDnsNamespaceError>;

    fn create_service(
        &self, 
        input: CreateServiceRequest
    ) -> RusotoFuture<CreateServiceResponse, CreateServiceError>;

    fn delete_namespace(
        &self, 
        input: DeleteNamespaceRequest
    ) -> RusotoFuture<DeleteNamespaceResponse, DeleteNamespaceError>;

    fn delete_service(
        &self, 
        input: DeleteServiceRequest
    ) -> RusotoFuture<DeleteServiceResponse, DeleteServiceError>;

    fn deregister_instance(
        &self, 
        input: DeregisterInstanceRequest
    ) -> RusotoFuture<DeregisterInstanceResponse, DeregisterInstanceError>;

    fn get_instance(
        &self, 
        input: GetInstanceRequest
    ) -> RusotoFuture<GetInstanceResponse, GetInstanceError>;

    fn get_instances_health_status(
        &self, 
        input: GetInstancesHealthStatusRequest
    ) -> RusotoFuture<GetInstancesHealthStatusResponse, GetInstancesHealthStatusError>;

    fn get_namespace(
        &self, 
        input: GetNamespaceRequest
    ) -> RusotoFuture<GetNamespaceResponse, GetNamespaceError>;

    fn get_operation(
        &self, 
        input: GetOperationRequest
    ) -> RusotoFuture<GetOperationResponse, GetOperationError>;

    fn get_service(
        &self, 
        input: GetServiceRequest
    ) -> RusotoFuture<GetServiceResponse, GetServiceError>;

    fn list_instances(
        &self, 
        input: ListInstancesRequest
    ) -> RusotoFuture<ListInstancesResponse, ListInstancesError>;

    fn list_namespaces(
        &self, 
        input: ListNamespacesRequest
    ) -> RusotoFuture<ListNamespacesResponse, ListNamespacesError>;

    fn list_operations(
        &self, 
        input: ListOperationsRequest
    ) -> RusotoFuture<ListOperationsResponse, ListOperationsError>;

    fn list_services(
        &self, 
        input: ListServicesRequest
    ) -> RusotoFuture<ListServicesResponse, ListServicesError>;

    fn register_instance(
        &self, 
        input: RegisterInstanceRequest
    ) -> RusotoFuture<RegisterInstanceResponse, RegisterInstanceError>;

    fn update_instance_custom_health_status(
        &self, 
        input: UpdateInstanceCustomHealthStatusRequest
    ) -> RusotoFuture<(), UpdateInstanceCustomHealthStatusError>;

    fn update_service(
        &self, 
        input: UpdateServiceRequest
    ) -> RusotoFuture<UpdateServiceResponse, UpdateServiceError>;
}

Trait representing the capabilities of the ServiceDiscovery API. ServiceDiscovery clients implement this trait.

Required Methods

fn create_private_dns_namespace(
&self,
input: CreatePrivateDnsNamespaceRequest
) -> RusotoFuture<CreatePrivateDnsNamespaceResponse, CreatePrivateDnsNamespaceError>

Creates a private namespace based on DNS, which will be visible only inside a specified Amazon VPC. The namespace defines your service naming scheme. For example, if you name your namespace example.com and name your service backend, the resulting DNS name for the service will be backend.example.com. For the current limit on the number of namespaces that you can create using the same AWS account, see Limits on Auto Naming in the Route 53 Developer Guide.

fn create_public_dns_namespace(
&self,
input: CreatePublicDnsNamespaceRequest
) -> RusotoFuture<CreatePublicDnsNamespaceResponse, CreatePublicDnsNamespaceError>

Creates a public namespace based on DNS, which will be visible on the internet. The namespace defines your service naming scheme. For example, if you name your namespace example.com and name your service backend, the resulting DNS name for the service will be backend.example.com. For the current limit on the number of namespaces that you can create using the same AWS account, see Limits on Auto Naming in the Route 53 Developer Guide.

fn create_service(
&self,
input: CreateServiceRequest
) -> RusotoFuture<CreateServiceResponse, CreateServiceError>

Creates a service, which defines the configuration for the following entities:

Up to three records (A, AAAA, and SRV) or one CNAME record

Optionally, a health check

After you create the service, you can submit a RegisterInstance request, and Amazon Route 53 uses the values in the configuration to create the specified entities.

For the current limit on the number of instances that you can register using the same namespace and using the same service, see Limits on Auto Naming in the Route 53 Developer Guide.

fn delete_namespace(
&self,
input: DeleteNamespaceRequest
) -> RusotoFuture<DeleteNamespaceResponse, DeleteNamespaceError>

Deletes a namespace from the current account. If the namespace still contains one or more services, the request fails.

fn delete_service(
&self,
input: DeleteServiceRequest
) -> RusotoFuture<DeleteServiceResponse, DeleteServiceError>

Deletes a specified service. If the service still contains one or more registered instances, the request fails.

fn deregister_instance(
&self,
input: DeregisterInstanceRequest
) -> RusotoFuture<DeregisterInstanceResponse, DeregisterInstanceError>

Deletes the records and the health check, if any, that Amazon Route 53 created for the specified instance.

fn get_instance(
&self,
input: GetInstanceRequest
) -> RusotoFuture<GetInstanceResponse, GetInstanceError>

Gets information about a specified instance.

fn get_instances_health_status(
&self,
input: GetInstancesHealthStatusRequest
) -> RusotoFuture<GetInstancesHealthStatusResponse, GetInstancesHealthStatusError>

Gets the current health status (Healthy, Unhealthy, or Unknown) of one or more instances that are associated with a specified service.

There is a brief delay between when you register an instance and when the health status for the instance is available.

fn get_namespace(
&self,
input: GetNamespaceRequest
) -> RusotoFuture<GetNamespaceResponse, GetNamespaceError>

Gets information about a namespace.

fn get_operation(
&self,
input: GetOperationRequest
) -> RusotoFuture<GetOperationResponse, GetOperationError>

Gets information about any operation that returns an operation ID in the response, such as a CreateService request.

To get a list of operations that match specified criteria, see ListOperations.

fn get_service(
&self,
input: GetServiceRequest
) -> RusotoFuture<GetServiceResponse, GetServiceError>

Gets the settings for a specified service.

fn list_instances(
&self,
input: ListInstancesRequest
) -> RusotoFuture<ListInstancesResponse, ListInstancesError>

Lists summary information about the instances that you registered by using a specified service.

fn list_namespaces(
&self,
input: ListNamespacesRequest
) -> RusotoFuture<ListNamespacesResponse, ListNamespacesError>

Lists summary information about the namespaces that were created by the current AWS account.

fn list_operations(
&self,
input: ListOperationsRequest
) -> RusotoFuture<ListOperationsResponse, ListOperationsError>

Lists operations that match the criteria that you specify.

fn list_services(
&self,
input: ListServicesRequest
) -> RusotoFuture<ListServicesResponse, ListServicesError>

Lists summary information for all the services that are associated with one or more specified namespaces.

fn register_instance(
&self,
input: RegisterInstanceRequest
) -> RusotoFuture<RegisterInstanceResponse, RegisterInstanceError>

Creates or updates one or more records and optionally a health check based on the settings in a specified service. When you submit a RegisterInstance request, Amazon Route 53 does the following:

For each DNS record that you define in the service specified by ServiceId, creates or updates a record in the hosted zone that is associated with the corresponding namespace

If the service includes HealthCheckConfig, creates or updates a health check based on the settings in the health check configuration

Associates the health check, if any, with each of the records

One RegisterInstance request must complete before you can submit another request and specify the same service ID and instance ID.

For more information, see CreateService.

When Route 53 receives a DNS query for the specified DNS name, it returns the applicable value:

If the health check is healthy: returns all the records

If the health check is unhealthy: returns the applicable value for the last healthy instance

If you didn't specify a health check configuration: returns all the records

For the current limit on the number of instances that you can register using the same namespace and using the same service, see Limits on Auto Naming in the Route 53 Developer Guide.

fn update_instance_custom_health_status(
&self,
input: UpdateInstanceCustomHealthStatusRequest
) -> RusotoFuture<(), UpdateInstanceCustomHealthStatusError>

fn update_service(
&self,
input: UpdateServiceRequest
) -> RusotoFuture<UpdateServiceResponse, UpdateServiceError>

Submits a request to perform the following operations:

Add or delete DnsRecords configurations

Update the TTL setting for existing DnsRecords configurations

Add, update, or delete HealthCheckConfig for a specified service

You must specify all DnsRecords configurations (and, optionally, HealthCheckConfig) that you want to appear in the updated service. Any current configurations that don't appear in an UpdateService request are deleted.

When you update the TTL setting for a service, Amazon Route 53 also updates the corresponding settings in all the records and health checks that were created by using the specified service.